Add missing python module import.
It's possible to not have python declared in a config.jam before trying to run the status tests. But the status tests add a python run target which requires python to be configured. This adds the import to at least allow for creating the default targets. And avoid hard errors.