Use either modular or not instal/stage targets for buildable libraries.
This account for the declared top-level install and stage targets to refer to either the modular /boost/library//stage|install targets. Or to refer to the non-modular libs/library/build//stage|install. This allows for the modular changes to be applied at random and the stage|install to still work with a mixed set of changes.